# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odoo Cafe POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ── USERS ───────────────────────────────────────────────────────────────────
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── PRODUCTS ─────────────────────────────────────────────────────────────────
...
  /floors:
    get:
      operationId: listFloors
      tags: [floors]
      summary: List floors with tables
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FloorWithTables"
    post:
      operationId: createFloor
      tags: [floors]
      summary: Create floor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
  /floors/{id}:
    patch:
      operationId: updateFloor
      tags: [floors]
      summary: Update floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
    delete:
      operationId: deleteFloor
      tags: [floors]
      summary: Delete floor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── TABLES ───────────────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Invalid coupon

  # ── CUSTOMERS ─────────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── ORDERS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SignupInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── User ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, active]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [User, Employee]
        active:
          type: boolean
...
    Category:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
    CategoryInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        color:
          type: string

    # ── Product ───────────────────────────────────────────────────────────────
...
    Floor:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    FloorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    FloorUpdate:
      type: object
      properties:
        name:
          type: string
    FloorWithTables:
      type: object
      required: [id, name, tables]
      properties:
        id:
          type: integer
        name:
          type: string
        tables:
          type: array
          items:
            $ref: "#/components/schemas/Table"

    # ── Table ─────────────────────────────────────────────────────────────────
    Table:
      type: object
      required: [id, floorId, number, seats, active]
      properties:
        id:
          type: integer
        floorId:
          type: integer
        floorName:
          type: ["string", "null"]
        number:
          type: integer
        seats:
          type: integer
        active:
          type: boolean
        hasActiveOrder:
          type: boolean
...
    Coupon:
      type: object
      required: [id, name, type, discountKind, discountValue, active]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [Coupon, Promotion]
        code:
          type: ["string", "null"]
        discountKind:
          type: string
          enum: [percent, amount]
        discountValue:
          type: number
        active:
          type: boolean
        apply:
          type: ["string", "null"]
          enum: [Product, Order, null]
        productId:
          type: ["integer", "null"]
        minQty:
          type: ["integer", "null"]
        minOrderAmount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
    CouponInput:
      type: object
      required: [name, type, discountKind, discountValue]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Coupon, Promotion]
        code:
          type: string
        discountKind:
          type: string
          enum: [percent, amount]
        discountValue:
          type: number
        active:
          type: boolean
        apply:
          type: string
          enum: [Product, Order]
        productId:
          type: integer
        minQty:
          type: integer
        minOrderAmount:
          type: number
        description:
          type: string
    CouponUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Coupon, Promotion]
        code:
          type: string
        discountKind:
          type: string
          enum: [percent, amount]
        discountValue:
          type: number
        active:
          type: boolean
        apply:
          type: string
          enum: [Product, Order]
        productId:
          type: integer
        minQty:
          type: integer
        minOrderAmount:
          type: number
        description:
          type: string
    CouponValidateInput:
      type: object
      required: [code]
      properties:
        code:
          type: string

    # ── Customer ──────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, email, phone]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string

    # ── Order ─────────────────────────────────────────────────────────────────
...